+++ Source/Starfish.Service/Application/Commands/Team/TeamMemberEditCommand.cs
@@ -4,16 +4,16 @@ namespace Nerosoft.Starfish.Application;
 
 public class TeamMemberEditCommand : Command
 {
-	public TeamMemberEditCommand(int teamId, List<int> userIds, string type)
+	public TeamMemberEditCommand(long teamId, List<long> userIds, string type)
 	{
 		TeamId = teamId;
 		UserIds = userIds;
 		Type = type;
 	}
 
-	public int TeamId { get; set; }
+	public long TeamId { get; set; }
 
-	public List<int> UserIds { get; set; }
+	public List<long> UserIds { get; set; }
 
 	public string Type { get; set; }
 }
