+++ Source/Starfish.Service/Application/Contracts/ITeamApplicationService.cs
@@ -9,17 +9,17 @@ public interface ITeamApplicationService : IApplicationService
 
 	Task<int> CountAsync(TeamCriteria criteria, CancellationToken cancellationToken = default);
 
-	Task<TeamDetailDto> GetAsync(int id, CancellationToken cancellationToken = default);
+	Task<TeamDetailDto> GetAsync(long id, CancellationToken cancellationToken = default);
 	
-	Task<int> CreateAsync(TeamEditDto data, CancellationToken cancellationToken = default);
+	Task<long> CreateAsync(TeamEditDto data, CancellationToken cancellationToken = default);
 
-	Task UpdateAsync(int id, TeamEditDto data, CancellationToken cancellationToken = default);
+	Task UpdateAsync(long id, TeamEditDto data, CancellationToken cancellationToken = default);
 
-	Task<List<TeamMemberDto>> QueryMembersAsync(int id, CancellationToken cancellationToken = default);
+	Task<List<TeamMemberDto>> QueryMembersAsync(long id, CancellationToken cancellationToken = default);
 
-	Task AppendMembersAsync(int id, List<int> userIds, CancellationToken cancellationToken = default);
+	Task AppendMembersAsync(long id, List<long> userIds, CancellationToken cancellationToken = default);
 
-	Task RemoveMembersAsync(int id, List<int> userIds, CancellationToken cancellationToken = default);
+	Task RemoveMembersAsync(long id, List<long> userIds, CancellationToken cancellationToken = default);
 
-	Task QuitAsync(int id, CancellationToken cancellationToken = default);
+	Task QuitAsync(long id, CancellationToken cancellationToken = default);
 }
