+++ Source/Starfish.Service/Domain/Exceptions/TeamNotFoundException.cs
@@ -2,7 +2,7 @@
 
 public class TeamNotFoundException : NotFoundException
 {
-	public TeamNotFoundException(int id)
+	public TeamNotFoundException(long id)
 		: base(string.Format(Resources.IDS_ERROR_TEAM_NOT_EXISTS, id))
 	{
 	}
