+++ Source/Starfish.Service/Domain/Repositories/ITeamRepository.cs
@@ -3,7 +3,7 @@ using Nerosoft.Starfish.Service;
 
 namespace Nerosoft.Starfish.Domain;
 
-public interface ITeamRepository : IBaseRepository<DataContext, Team, int>
+public interface ITeamRepository : IBaseRepository<DataContext, Team, long>
 {
 	/// <summary>
 	/// 查询指定用户所属团队
@@ -11,7 +11,7 @@ public interface ITeamRepository : IBaseRepository<DataContext, Team, int>
 	/// <param name="userId"></param>
 	/// <param name="cancellationToken"></param>
 	/// <returns></returns>
-	Task<List<Team>> GetTeamsOfUserAsync(int userId, CancellationToken cancellationToken = default);
+	Task<List<Team>> GetTeamsOfUserAsync(long userId, CancellationToken cancellationToken = default);
 
-	Task<List<TeamMember>> GetMembersAsync(int id, CancellationToken cancellationToken = default);
+	Task<List<TeamMember>> GetMembersAsync(long id, CancellationToken cancellationToken = default);
 }
