+++ Source/Starfish.Service/Domain/Repositories/ITokenRepository.cs
@@ -6,6 +6,6 @@ namespace Nerosoft.Starfish.Domain;
 /// <summary>
 /// 令牌仓储接口
 /// </summary>
-public interface ITokenRepository : IBaseRepository<DataContext, Token, int>
+public interface ITokenRepository : IBaseRepository<DataContext, Token, long>
 {
 }
