+++ Source/Starfish.Service/Domain/Repositories/IUserRepository.cs
@@ -6,7 +6,7 @@ namespace Nerosoft.Starfish.Domain;
 /// <summary>
 /// 用户仓储接口
 /// </summary>
-public interface IUserRepository : IBaseRepository<DataContext, User, int>
+public interface IUserRepository : IBaseRepository<DataContext, User, long>
 {
 	/// <summary>
 	/// 根据用户名查询用户
@@ -32,7 +32,7 @@ public interface IUserRepository : IBaseRepository<DataContext, User, int>
 	/// <param name="ignoreId"></param>
 	/// <param name="cancellationToken"></param>
 	/// <returns></returns>
-	Task<bool> CheckEmailExistsAsync(string email, int ignoreId, CancellationToken cancellationToken = default);
+	Task<bool> CheckEmailExistsAsync(string email, long ignoreId, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// 检查手机号是否存在
@@ -41,5 +41,5 @@ public interface IUserRepository : IBaseRepository<DataContext, User, int>
 	/// <param name="ignoreId"></param>
 	/// <param name="cancellationToken"></param>
 	/// <returns></returns>
-	Task<bool> CheckPhoneExistsAsync(string phone, int ignoreId, CancellationToken cancellationToken = default);
+	Task<bool> CheckPhoneExistsAsync(string phone, long ignoreId, CancellationToken cancellationToken = default);
 }
