+++ Source/Starfish.Service/Repository/Repositories/TeamRepository.cs
@@ -5,14 +5,14 @@ using Nerosoft.Starfish.Service;
 
 namespace Nerosoft.Starfish.Repository;
 
-public class TeamRepository : BaseRepository<DataContext, Team, int>, ITeamRepository
+public class TeamRepository : BaseRepository<DataContext, Team, long>, ITeamRepository
 {
 	public TeamRepository(IContextProvider provider)
 		: base(provider)
 	{
 	}
 
-	public Task<List<Team>> GetTeamsOfUserAsync(int userId, CancellationToken cancellationToken = default)
+	public Task<List<Team>> GetTeamsOfUserAsync(long userId, CancellationToken cancellationToken = default)
 	{
 		var memberSet = Context.Set<TeamMember>();
 		var teamSet = Context.Set<Team>();
@@ -23,7 +23,7 @@ public class TeamRepository : BaseRepository<DataContext, Team, int>, ITeamRepos
 		return query.ToListAsync(cancellationToken);
 	}
 
-	public Task<List<TeamMember>> GetMembersAsync(int id, CancellationToken cancellationToken = default)
+	public Task<List<TeamMember>> GetMembersAsync(long id, CancellationToken cancellationToken = default)
 	{
 		var query = Context.Set<TeamMember>()
 		                   .Include(t => t.User)
