+++ Source/Starfish.Service/Repository/Specifications/TeamSpecification.cs
@@ -5,7 +5,7 @@ namespace Nerosoft.Starfish.Repository;
 
 internal static class TeamSpecification
 {
-	public static Specification<Team> IdEquals(int id)
+	public static Specification<Team> IdEquals(long id)
 	{
 		return new DirectSpecification<Team>(t => t.Id == id);
 	}
@@ -45,7 +45,7 @@ internal static class TeamSpecification
 		return new CompositeSpecification<Team>(PredicateOperator.OrElse, specifications);
 	}
 
-	public static Specification<Team> HasMember(int userId)
+	public static Specification<Team> HasMember(long userId)
 	{
 		return new DirectSpecification<Team>(t => t.Members.Any(m => m.UserId == userId));
 	}
