+++ Source/Starfish.Service/UseCases/Team/TeamMemberAppendUseCase.cs
@@ -6,7 +6,7 @@ namespace Nerosoft.Starfish.UseCases;
 
 public interface ITeamMemberAppendUseCase : INonOutputUseCase<TeamMemberAppendInput>;
 
-public record TeamMemberAppendInput(int Id, List<int> UserIds) : IUseCaseInput;
+public record TeamMemberAppendInput(long Id, List<long> UserIds) : IUseCaseInput;
 
 public class TeamMemberAppendUseCase : ITeamMemberAppendUseCase
 {
