+++ Source/Starfish.Transit/Team/TeamDetailDto.cs
@@ -8,7 +8,7 @@ public class TeamDetailDto
 	/// <summary>
 	/// Id
 	/// </summary>
-	public int Id { get; set; }
+	public long Id { get; set; }
 
 	/// <summary>
 	/// 别名
@@ -28,7 +28,7 @@ public class TeamDetailDto
 	/// <summary>
 	/// 团队所有者Id
 	/// </summary>
-	public int OwnerId { get; set; }
+	public long OwnerId { get; set; }
 
 	/// <summary>
 	/// 成员数量
